Extract shared wrapper edge client and document KMS path
The langchain, crewai, and autogen wrappers each carried ~100 lines of
identical HTTP plumbing for the MCP, AWI, and balance endpoints. Move
that surface into a single `B2AEdgeClient` base in `b2a_sdk`, have each
wrapper inherit from it, and add `b2a-sdk` as a wrapper dependency so
contract changes happen in one place.

Add `docs/key-management.md` describing the current env-var-based loader
in `app/services/signing_keys.py`, why it is acceptable for MVP, the
intended external-KMS production target (AWS KMS / GCP KMS / Vault
Transit), the specific files that change, the rotation flow, and the
compromise-response procedure. This is the answer to "where do the
private keys live?" that the SECURITY_LIMITATIONS document defers.
