+++ wrappers/autogen-agent-middleware/pyproject.toml
@@ -10,6 +10,7 @@ readme = "README.md"
 license = "MIT"
 requires-python = ">=3.11"
 dependencies = [
+    "b2a-sdk>=0.2.0",
     "autogen-agentchat>=0.2.0",
     "httpx>=0.25.0",
 ]
