+++ wrappers/langchain-agent-middleware/pyproject.toml
@@ -10,6 +10,7 @@ readme = "README.md"
 license = "MIT"
 requires-python = ">=3.11"
 dependencies = [
+    "b2a-sdk>=0.2.0",
     "langchain>=0.1.0",
     "langgraph>=0.0.20",
     "httpx>=0.25.0",
